GEOMETRY-144
Delete unnecessary abstract parent class and move tests to
ConvexHullBuilderTest
